translator_requirements.txt
funcy==1.16
pika==1.3.0
elasticsearch==8.12.1
pymongo==4.1.1
retry==0.9.2
redis==5.0.4
requests==2.31.0
isodate==0.6.1
shapely==2.0.5
ai4bharat-transliteration==1.1.3